+++ src/main/java/com/brachy84/mechtech/comon/items/MTMetaItem.java
@@ -1,11 +1,10 @@
 package com.brachy84.mechtech.comon.items;
 
-import com.brachy84.mechtech.api.armor.ModularArmor;
-import com.brachy84.mechtech.comon.items.behavior.DataStickBehavior;
+import com.brachy84.mechtech.api.armor.Modules;
 import gregtech.api.items.metaitem.StandardMetaItem;
 import gregtech.common.items.MetaItems;
 
-import static com.brachy84.mechtech.comon.items.MTMetaItems.*;
+import static com.brachy84.mechtech.comon.items.MTMetaItems.WIRELESS_RECEIVER;
 
 public class MTMetaItem extends StandardMetaItem {
 
@@ -19,10 +18,10 @@ public class MTMetaItem extends StandardMetaItem {
         //NIGHTVISION_MODULE = addItem(100, "module.night_vision").addComponents(ModularArmor.Modules.nightVision);
         //WIRELESS_BINDER = addItem(2001, "wireless_binder");
 
-        MetaItems.NIGHTVISION_GOGGLES.addComponents(ModularArmor.Modules.nightVision);
-        MetaItems.COVER_SOLAR_PANEL_LV.addComponents(ModularArmor.Modules.solarGen1);
-        MetaItems.COVER_SOLAR_PANEL_MV.addComponents(ModularArmor.Modules.solarGen2);
-        MetaItems.COVER_SOLAR_PANEL_HV.addComponents(ModularArmor.Modules.solarGen3);
+        MetaItems.NIGHTVISION_GOGGLES.addComponents(Modules.nightVision);
+        MetaItems.COVER_SOLAR_PANEL_LV.addComponents(Modules.solarGen1);
+        MetaItems.COVER_SOLAR_PANEL_MV.addComponents(Modules.solarGen2);
+        MetaItems.COVER_SOLAR_PANEL_HV.addComponents(Modules.solarGen3);
         //MetaItems.TOOL_DATA_STICK.addComponents(new DataStickBehavior());
     }
 }
